+++ src/main/java/org/team1540/robot2019/OI.java
@@ -8,6 +8,7 @@ import edu.wpi.first.wpilibj.buttons.Button;
 import edu.wpi.first.wpilibj.buttons.JoystickButton;
 import edu.wpi.first.wpilibj.command.Command;
 import org.apache.log4j.Logger;
+import org.team1540.robot2019.commands.auto.PurePursuitThenPointToVisionTarget;
 import org.team1540.robot2019.commands.cargo.EjectThenDown;
 import org.team1540.robot2019.commands.cargo.FloorIntake;
 import org.team1540.robot2019.commands.cargo.LoadingStationIntake;
@@ -20,7 +21,6 @@ import org.team1540.robot2019.commands.hatch.GetHatchFloor;
 import org.team1540.robot2019.commands.hatch.GrabHatch;
 import org.team1540.robot2019.commands.hatch.PlaceHatchThenDown;
 import org.team1540.robot2019.commands.hatch.RetractHatch;
-import org.team1540.robot2019.drivecontrol.commands.PurePursuitThenPointToVisionTarget;
 import org.team1540.rooster.Utilities;
 import org.team1540.rooster.triggers.AxisButton;
 import org.team1540.rooster.triggers.DPadAxis;
