+++ src/main/java/org/team1540/robot2019/odometry/tankdrive/TankDriveOdometryRunnable.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2019.odometry;
+package org.team1540.robot2019.odometry.tankdrive;
 
 import edu.wpi.first.wpilibj.Notifier;
 import java.util.function.DoubleSupplier;
