+++ src/main/java/org/team1540/robot2019/odometry/tankdrive/TankDriveOdometryUtils.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2019.odometry;
+package org.team1540.robot2019.odometry.tankdrive;
 
 
 import org.team1540.robot2019.datastructures.threed.Transform3D;
